Only service owner now can edit and delete it

+++ FromLocalsToLocals/Controllers/VendorsController.cs
@@ -6,25 +6,34 @@ using FromLocalsToLocals.Database;
 using FromLocalsToLocals.Models;
 using SuppLocals;
 using Microsoft.AspNetCore.Authorization;
+using FromLocalsToLocals.ViewModels;
+using Microsoft.AspNetCore.Identity;
+using System.Security.Policy;
 
 namespace FromLocalsToLocals.Controllers
 {
+    [Authorize]
     public class VendorsController : Controller
     {
         private readonly AppDbContext _context;
+        private readonly UserManager<AppUser> _userManager;
 
-        public VendorsController(AppDbContext context)
+        public VendorsController(AppDbContext context, UserManager<AppUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
-        // GET: Vendors
-        public async Task<IActionResult> Index()
+
+        [HttpGet]
+        [AllowAnonymous]
+        public async Task<IActionResult> AllVendors()
         {
             return View(await _context.Vendors.ToListAsync());
         }
 
-        // GET: Vendors/Details/5
+        [HttpGet]
+        [AllowAnonymous]
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -42,69 +51,69 @@ namespace FromLocalsToLocals.Controllers
             return View(vendor);
         }
 
-        // GET: Vendors/Create
-        [Authorize]
+        [HttpGet]
         public IActionResult Create()
         {
             return View();
         }
 
-        // POST: Vendors/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
-        [Authorize]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create(Vendor vendor)
+        public async Task<IActionResult> Create(Vendor model)
         {
             if (ModelState.IsValid)
             {
-                var latLng = await MapMethods.ConvertAddressToLocationAsync(vendor.Address);
+                var latLng = await MapMethods.ConvertAddressToLocationAsync(model.Address);
 
                 if (latLng != null)
                 {
-                    //Temporary until login system will be created
-                    vendor.UserID = 0;
 
-                    vendor.Latitude = latLng.Item1;
-                    vendor.Longitude = latLng.Item2;
-
-                    _context.Vendors.Add(vendor);
+                    model.UserID = _userManager.GetUserId(User);
+                    model.Latitude = latLng.Item1;
+                    model.Longitude = latLng.Item2;
+                    
+                    _context.Vendors.Add(model);
                     _context.SaveChanges();
 
-                    return RedirectToAction("Index");
-
+                    return RedirectToAction("AllVendors");
                 }
             }
 
             //Somehow we should inform the client-side that probably address is not recognizable
-            return View(vendor);
+            return View(model);
         }
 
-        // GET: Vendors/Edit/5
+        [HttpGet]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
-
+            
             var vendor = await _context.Vendors.FindAsync(id);
             if (vendor == null)
             {
                 return NotFound();
             }
+
+            if (!ValidUser(vendor.UserID))
+            {
+                return NotFound();
+            }
+
             return View(vendor);
         }
 
-        // POST: Vendors/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, Vendor vendor)
+        public async Task<IActionResult> Edit(int id, Vendor model)
         {
-            if (id != vendor.ID)
+            if (id != model.ID)
+            {
+                return NotFound();
+            }
+            if (model.UserID != _userManager.GetUserId(User))
             {
                 return NotFound();
             }
@@ -113,23 +122,24 @@ namespace FromLocalsToLocals.Controllers
             {
                 try
                 {
-                    var latLng = await MapMethods.ConvertAddressToLocationAsync(vendor.Address);
+                    var latLng = await MapMethods.ConvertAddressToLocationAsync(model.Address);
 
                     if (latLng == null)
                     {
                         //Somehow we should inform the client-side that probably address is not recognizable
-                        return View(vendor);
+                        return View(model);
                     }
-                    vendor.Latitude = latLng.Item1;
-                    vendor.Longitude = latLng.Item2;
 
-                    _context.Update(vendor);
+                    model.Latitude = latLng.Item1;
+                    model.Longitude = latLng.Item2;
+
+                    _context.Update(model);
                     await _context.SaveChangesAsync();
 
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!VendorExists(vendor.ID))
+                    if (!VendorExists(model.ID))
                     {
                         return NotFound();
                     }
@@ -138,12 +148,12 @@ namespace FromLocalsToLocals.Controllers
                         throw;
                     }
                 }
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction(nameof(AllVendors));
             }
-            return View(vendor);
+            return View(model);
         }
 
-        // GET: Vendors/Delete/5
+        [HttpGet]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -158,23 +168,38 @@ namespace FromLocalsToLocals.Controllers
                 return NotFound();
             }
 
+            if (!ValidUser(vendor.UserID))
+            {
+                return NotFound();
+            }
+
             return View(vendor);
         }
 
-        // POST: Vendors/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var vendor = await _context.Vendors.FindAsync(id);
+
+            if (!ValidUser(vendor.UserID))
+            {
+                return NotFound();
+            }
+
             _context.Vendors.Remove(vendor);
             await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction(nameof(AllVendors));
         }
 
         private bool VendorExists(int id)
         {
             return _context.Vendors.Any(e => e.ID == id);
+        }
+
+        private bool ValidUser(string id)
+        {
+            return id == _userManager.GetUserId(User);
         }
     }
 }
